Fix api tests
